refactor: pg사 관련 변수 PgServiceImpl가 갖고, 이 변수를 사용하는 모든 메서드를 구현하도록 코드 수정
