+++ src/main/java/com/matsinger/barofishserver/domain/payment/portone/application/PgServiceImpl.java
@@ -1,7 +1,11 @@
 package com.matsinger.barofishserver.domain.payment.portone.application;
 
+import com.matsinger.barofishserver.domain.order.domain.BankCode;
 import com.matsinger.barofishserver.domain.payment.dto.CheckValidCardRes;
 import com.matsinger.barofishserver.domain.payment.dto.IamPortCertificationRes;
+import com.matsinger.barofishserver.domain.payment.portone.dto.PortOneAccessKeyRequest;
+import com.matsinger.barofishserver.domain.payment.portone.dto.PortOneAccessKeyResponse;
+import com.matsinger.barofishserver.domain.payment.portone.dto.PortOneVbankHolderResponse;
 import com.matsinger.barofishserver.global.exception.BusinessException;
 import com.siot.IamportRestClient.IamportClient;
 import com.siot.IamportRestClient.request.AgainPaymentData;
@@ -15,14 +19,19 @@ import jakarta.annotation.PostConstruct;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Profile;
+import org.springframework.http.*;
 import org.springframework.stereotype.Service;
+import org.springframework.web.client.HttpClientErrorException;
+import org.springframework.web.client.RestTemplate;
+import org.springframework.web.util.UriComponentsBuilder;
 
+import java.net.URI;
 import java.util.stream.Stream;
 
 @Slf4j
 @Service
 @Profile("!local")
-public class PortOneCallbackService implements PgService {
+public class PgServiceImpl implements PgService {
 
     @Value("${iamport.credentials.apiKey}")
     private String accessKey;
@@ -30,8 +39,16 @@ public class PortOneCallbackService implements PgService {
     @Value("${iamport.credentials.secretKey}")
     private String secretKey;
 
+    @Value("${iamport.credentials.keyin-pg}")
+    public String keyinPg;
+
+    @Value("${iamport.credentials.mid}")
+    public String mid;
+
     @Value("${iamport.webhook.url}")
     private String webhookUrl;
+    private String portOneUrl = "https://api.iamport.kr/";
+    private String accessToken;
 
     private IamportClient iamportClient;
 
@@ -115,6 +132,7 @@ public class PortOneCallbackService implements PgService {
 
     @Override
     public CheckValidCardRes processPayment(String uid, BillingCustomerData customerData) {
+        customerData.setPg(keyinPg);
         try {
             IamportResponse<BillingCustomer> portoneResponse = iamportClient.postBillingCustomer(uid, customerData);
             BillingCustomer billingCustomer = portoneResponse.getResponse();
@@ -130,5 +148,93 @@ public class PortOneCallbackService implements PgService {
         } catch (Exception e) {
             throw new BusinessException("결제 요청에 실패했습니다.");
         }
+    }
+
+    @Override
+    public String checkVbankAccountVerification(BankCode bankCode, String bankNum, String holderName) {
+        URI uri = UriComponentsBuilder
+                .fromUriString(portOneUrl)
+                .path("vbanks/holder")
+                .queryParam("bank_code", bankCode.getCode())
+                .queryParam("bank_num", bankNum)
+                .encode()
+                .build()
+                .toUri();
+
+        generateAccessToken();
+        return sendPortOneAccountInfo(uri, holderName);
+    }
+
+    private String sendPortOneAccountInfo(URI uri, String holderName) {
+        RestTemplate restTemplate = new RestTemplate();
+        HttpHeaders headers = new HttpHeaders();
+        headers.setContentType(MediaType.APPLICATION_JSON);
+        headers.setBearerAuth(accessToken);
+        HttpEntity request = new HttpEntity(headers);
+
+        String errorMessage = null;
+        int trialCount = 0;
+        try {
+            ResponseEntity<PortOneVbankHolderResponse> responseEntity = checkBankAccountValid(uri, restTemplate, request);
+            String bankHolderResponse = responseEntity.getBody().getResponse().getBank_holder();
+            if (!bankHolderResponse.equals(holderName)) {
+                errorMessage = "계좌 소유주 명이 일치하지 않습니다.";
+            }
+        } catch (HttpClientErrorException e) {
+            trialCount++;
+            if (trialCount < 2) {
+                if (e.getStatusCode().equals(HttpStatus.UNAUTHORIZED)) {
+                    generateAccessToken();
+                    ResponseEntity<PortOneVbankHolderResponse> responseEntity = checkBankAccountValid(uri, restTemplate, request);
+                    String bankHolderResponse = responseEntity.getBody().getResponse().getBank_holder();
+                    if (!bankHolderResponse.equals(holderName)) {
+                        errorMessage = "계좌 소유주 명이 일치하지 않습니다.";
+                    }
+                }
+            }
+
+            if (e.getStatusCode().equals(HttpStatus.BAD_REQUEST)) {
+                errorMessage = "은행, 또는 계좌번호를 입력해주세요.";
+            }
+            if (e.getStatusCode().equals(HttpStatus.NOT_FOUND)) {
+                errorMessage = "계좌 번호가 일치하지 않습니다.";
+            }
+        }
+        return errorMessage;
+    }
+
+    @Override
+    public ResponseEntity<PortOneVbankHolderResponse> checkBankAccountValid(URI uri, RestTemplate restTemplate, HttpEntity request) {
+        ResponseEntity<PortOneVbankHolderResponse> responseEntity = restTemplate.exchange(
+                uri,
+                HttpMethod.GET,
+                request,
+                PortOneVbankHolderResponse.class
+        );
+        return responseEntity;
+    }
+
+    public void generateAccessToken() {
+        URI uri = UriComponentsBuilder
+                .fromUriString(portOneUrl)
+                .path("users/getToken")
+                .encode()
+                .build()
+                .toUri();
+        RestTemplate restTemplate = new RestTemplate();
+
+        HttpHeaders headers = new HttpHeaders();
+        headers.setContentType(MediaType.APPLICATION_JSON);
+        PortOneAccessKeyRequest requestDto = new PortOneAccessKeyRequest(accessKey, secretKey);
+        HttpEntity<PortOneAccessKeyRequest> request = new HttpEntity<>(requestDto, headers);
+
+        ResponseEntity<PortOneAccessKeyResponse> responseEntity = restTemplate.exchange(
+                uri,
+                HttpMethod.POST,
+                request,
+                PortOneAccessKeyResponse.class
+        );
+
+        this.accessToken = responseEntity.getBody().getResponse().getAccess_token();
     }
 }
